Cleaning up + update readme
