+++ .gitignore
@@ -2,3 +2,4 @@ __pycache__
 gcp_key.json
 logs
 local_client.py
+*.pkl
